+++ packages/circuit-ui/components/CheckboxGroup/CheckboxGroup.tsx
@@ -35,6 +35,7 @@ import {
   isSufficientlyLabelled,
 } from '../../util/errors.js';
 import { isEmpty } from '../../util/helpers.js';
+import { idx } from '../../util/idx.js';
 
 import classes from './CheckboxGroup.module.css';
 
@@ -146,9 +147,10 @@ export const CheckboxGroup = forwardRef(
     ref: CheckboxGroupProps['ref'],
   ) => {
     const validationHintId = useId();
-    const descriptionIds = `${
-      descriptionId ? `${descriptionId} ` : ''
-    }${validationHintId}`;
+    const descriptionIds = idx(
+      descriptionId,
+      validationHint && validationHintId,
+    );
 
     if (
       process.env.NODE_ENV !== 'production' &&