+++ packages/circuit-ui/components/Pagination/components/PageList/PageList.tsx
@@ -38,8 +38,7 @@ export const PageList: FC<PageListProps> = ({
   className,
   ...props
 }: PageListProps) => (
-  // eslint-disable-next-line jsx-a11y/no-redundant-roles
-  <ol role="list" className={clsx(classes.base, className)} {...props}>
+  <ol className={clsx(classes.base, className)} {...props}>
     {pages.map((page) => {
       const isCurrent = currentPage === page;
       const label = pageLabel(page);