+++ packages/circuit-ui/components/SelectorGroup/SelectorGroup.tsx
@@ -34,6 +34,7 @@ import {
 } from '../Field/index.js';
 import { isEmpty } from '../../util/helpers.js';
 import { clsx } from '../../styles/clsx.js';
+import { idx } from '../../util/idx.js';
 
 import classes from './SelectorGroup.module.css';
 
@@ -150,9 +151,10 @@ export const SelectorGroup = forwardRef<
     const randomName = useId();
     const name = customName || randomName;
     const validationHintId = useId();
-    const descriptionIds = `${
-      descriptionId ? `${descriptionId} ` : ''
-    }${validationHintId}`;
+    const descriptionIds = idx(
+      descriptionId,
+      validationHint && validationHintId,
+    );
 
     if (
       process.env.NODE_ENV !== 'production' &&