+++ packages/circuit-ui/components/SideNavigation/components/DesktopNavigation/DesktopNavigation.tsx
@@ -13,8 +13,6 @@
  * limitations under the License.
  */
 
-/* eslint-disable jsx-a11y/no-redundant-roles */
-
 'use client';
 
 import { utilClasses } from '../../../../styles/utility.js';
@@ -85,7 +83,7 @@ export function DesktopNavigation({
         {skipNavigationHref && skipNavigationLabel && (
           <SkipLink href={skipNavigationHref}>{skipNavigationLabel}</SkipLink>
         )}
-        <ul role="list" className={classes.list}>
+        <ul className={classes.list}>
           {primaryLinks.map((link) => (
             <li key={link.label}>
               <PrimaryLink {...link} {...focusProps} />