+++ packages/circuit-ui/components/SideNavigation/components/SecondaryLinks/SecondaryLinks.tsx
@@ -13,8 +13,6 @@
  * limitations under the License.
  */
 
-/* eslint-disable jsx-a11y/no-redundant-roles */
-
 'use client';
 
 import { forwardRef } from 'react';
@@ -98,7 +96,7 @@ function SecondaryGroup({
           </Body>
         </Skeleton>
       )}
-      <ul role="list" className={classes.list}>
+      <ul className={classes.list}>
         {secondaryLinks.map((link) => (
           <SecondaryLink key={link.label} {...link} {...focusProps} />
         ))}
@@ -116,12 +114,7 @@ export const SecondaryLinks = forwardRef<HTMLUListElement, SecondaryLinksProps>(
   ({ secondaryGroups, className, ...props }, ref) => {
     const focusProps = useFocusList();
     return (
-      <ul
-        role="list"
-        ref={ref}
-        className={clsx(classes.list, className)}
-        {...props}
-      >
+      <ul ref={ref} className={clsx(classes.list, className)} {...props}>
         {secondaryGroups.map((group) => (
           <SecondaryGroup
             key={group.label}